Added event listener functions to both timer classes.

+++ WebWorkerTimer.js
@@ -1,5 +1,15 @@
-function WebWorkerTimer(audioContext) {
+/*
+ * params: {
+ *   id: String, 
+ *   tickInterval: Number (opt) (miliseconds)
+ * }
+ */
+function WebWorkerTimer(params) {
 	
+	var _self = this;
+	var _id = params.id;
+	var _tickInterval = params.tickInterval || 0.025*1000;
+	var _counterId = 0;
 	var _callbacks = {};
 	var _timerJS = URL.createObjectURL(
 		new Blob([
@@ -16,28 +26,60 @@ function WebWorkerTimer(audioContext) {
 	);
     var _currentTime = 0;
     var _worker;
-    var _startTimeStamp;
+    var _offset;
+    var _running = false;
 
 
-	this.on_tick = function(id, callback) {
-		_callbacks[id] = callback;
+	var _emit = function(evenType, data) {
+		for (var ci in _callbacks[evenType]) 
+			_callbacks[evenType][ci](data);
+	}
+
+	this.on = function(observerID, eventType, callback) {
+
+		if (!eventType || _callbacks[eventType]==undefined) 
+			throws "Unsupported event type";
+
+		if (observerID!=undefined && _callbacks[eventType][observerID]!=undefined) 
+			throws "Illegal modification of callback";
+
+		var __id = (observerID==undefined)? _id + "-associate-" + (_idCounter++) : observerID;
+		_callbacks[eventType][__id] = callback;
+		return __id;
+	}
+
+	this.off = function(observerID, eventType) {
+		if (!eventType || _callbacks[eventType]==undefined) 
+			throws "Unsupported event type";
+
+		delete _callbacks[eventType][observerID];
 	}
 
 	this.start = function() {
-		_worker = new Worker(_timerJS);
-		_startTimeStamp = (new Date()).getTime();
-		_worker.onmessage = function(data) {
-			var t0 = _currentTime;
-			var t1 = (new Date()).getTime() - _startTimeStamp;
-    		_currentTime += t1 - t0;
-			for (var ci in _callbacks) 
-				_callbacks[ci](_currentTime);
-		};
+		if (!_running) {
+			_worker = new Worker(_timerJS);
+			_offset = (new Date()).getTime();
+
+			_worker.onmessage = function(data) {
+				var t0 = _currentTime;
+				var t1 = (new Date()).getTime() - _offset;
+	    		_currentTime += t1 - t0;
+	    		_emit('tick', {id: _id, time: _currentTime});
+			};
+
+			_running = true;
+			_emit('start', {id: _id});
+			_worker.postMessage(_tickInterval);
+		}
 	}
 
 	this.stop = function() {
-		_worker.postMessage(0);
-		_worker.terminate();
+		if (_running) {
+			_worker.postMessage(0);
+			_worker.terminate();
+			_running = false;
+			_emit('stop', {id: _id});
+		}
 	}
 
 	this.time = function() {
@@ -46,5 +88,10 @@ function WebWorkerTimer(audioContext) {
 
 	this.reset = function() {
 		_currentTime = 0;
+		_emit('reset', {id: _id});
+	}
+
+	this.id = function() {
+		return _id;
 	}
 }
